Split up backend types
