--- backend/types.ts
@@ -1,37 +1,0 @@
-export type Role = "Administrator" | "Facilitator" | "Learner";
-
-export type Token = {
-  accessToken: string;
-  refreshToken: string;
-};
-
-export type UserDTO = {
-  id: string;
-  firstName: string;
-  lastName: string;
-  email: string;
-  role: Role;
-};
-
-export type CreateUserDTO = Omit<UserDTO, "id"> & { password: string };
-
-export type UpdateUserDTO = Omit<UserDTO, "id">;
-
-export type SignupUserDTO = Omit<CreateUserDTO, "role">;
-
-export type AuthDTO = Token & UserDTO;
-
-export type Letters = "A" | "B" | "C" | "D";
-
-export type NodemailerConfig = {
-  service: "gmail";
-  auth: {
-    type: "OAuth2";
-    user: string;
-    clientId: string;
-    clientSecret: string;
-    refreshToken: string;
-  };
-};
-
-export type SignupMethod = "PASSWORD" | "GOOGLE";
