+++ backend/types/userTypes.ts
@@ -1,0 +1,15 @@
+export type Role = "Administrator" | "Facilitator" | "Learner";
+
+export type UserDTO = {
+  id: string;
+  firstName: string;
+  lastName: string;
+  email: string;
+  role: Role;
+};
+
+export type CreateUserDTO = Omit<UserDTO, "id"> & { password: string };
+
+export type UpdateUserDTO = Omit<UserDTO, "id">;
+
+export type SignupUserDTO = Omit<CreateUserDTO, "role">;
